Rename "thrift" podspec to "box-thrift"
Giving ours a unique name makes it possible to properly add it add as a dependency to other podspecs. Otherwise, `pod spec lint` on the consuming pod would return build errors, because it would try to build with the latest public release of thrift. There's no way in a podspec to point it at a particular fork, you can only do that from a Podfile. This wasn't a problem until now, because we didn't need to publish a podspec of a repo that depends on this pod.